Clean up notebook test utils
