+++ packages/notebook/test/utils.ts
@@ -1,15 +1,13 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { Context } from '@jupyterlab/docregistry';
-import { Mock, NBTestUtils } from '@jupyterlab/testutils';
-import { UUID } from '@lumino/coreutils';
+import { Context, DocumentRegistry } from '@jupyterlab/docregistry';
+import { NBTestUtils } from '@jupyterlab/testutils';
 import {
   INotebookModel,
   Notebook,
-  NotebookModel,
-  NotebookModelFactory,
-  NotebookPanel
+  NotebookPanel,
+  NotebookWidgetFactory
 } from '../src';
 
 /**
@@ -20,29 +18,21 @@ import {
  * Create a default notebook content factory.
  */
 export function createNotebookFactory(): Notebook.IContentFactory {
-  return new Notebook.ContentFactory({
-    editorFactory: NBTestUtils.editorFactory
-  });
+  return NBTestUtils.createNotebookFactory();
 }
 
 /**
  * Create a default notebook panel content factory.
  */
 export function createNotebookPanelFactory(): NotebookPanel.IContentFactory {
-  return new NotebookPanel.ContentFactory({
-    editorFactory: NBTestUtils.editorFactory
-  });
+  return NBTestUtils.createNotebookPanelFactory();
 }
 
 /**
  * Create a notebook widget.
  */
 export function createNotebook(): Notebook {
-  return new Notebook({
-    rendermime: NBTestUtils.defaultRenderMime(),
-    contentFactory: createNotebookFactory(),
-    mimeTypeService: NBTestUtils.mimeTypeService
-  });
+  return NBTestUtils.createNotebook();
 }
 
 /**
@@ -51,19 +41,14 @@ export function createNotebook(): Notebook {
 export function createNotebookPanel(
   context: Context<INotebookModel>
 ): NotebookPanel {
-  return new NotebookPanel({
-    content: createNotebook(),
-    context
-  });
+  return NBTestUtils.createNotebookPanel(context);
 }
 
 /**
  * Populate a notebook with default content.
  */
 export function populateNotebook(notebook: Notebook): void {
-  const model = new NotebookModel();
-  model.fromJSON(NBTestUtils.DEFAULT_CONTENT);
-  notebook.model = model;
+  NBTestUtils.populateNotebook(notebook);
 }
 
 export const DEFAULT_CONTENT = NBTestUtils.DEFAULT_CONTENT;
@@ -77,27 +62,17 @@ export function defaultRenderMime(): any {
   return NBTestUtils.defaultRenderMime();
 }
 
+export function createNotebookWidgetFactory(
+  toolbarFactory?: (widget: NotebookPanel) => DocumentRegistry.IToolbarItem[]
+): NotebookWidgetFactory {
+  return NBTestUtils.createNotebookWidgetFactory(toolbarFactory);
+}
+
 /**
  * Create a context for a file.
  */
 export async function createMockContext(
   startKernel = false
 ): Promise<Context<INotebookModel>> {
-  const path = UUID.uuid4() + '.txt';
-  const manager = new Mock.ServiceManagerMock();
-  const factory = new NotebookModelFactory({});
-
-  const context = new Context({
-    manager,
-    factory,
-    path,
-    kernelPreference: {
-      shouldStart: startKernel,
-      canStart: startKernel,
-      autoStartDefault: startKernel
-    }
-  });
-  await context.initialize(true);
-  await context.sessionContext.initialize();
-  return context;
+  return NBTestUtils.createMockContext(startKernel);
 }
